Modify PartTwo.cpp to order codetable by ASCII values Add codetable.h for PartTwo data structures
