+++ codetable.txt
@@ -1,36 +1,36 @@
-o 0000
-r 0001
-e 001
-W 0100000
-; 0100001
-A 0100010
-v 0100011
-, 010010
-b 010011
-h 0101
-l 0110
-y 01110
-i 01111
-t 1000
-n 10010
-. 1001100
-k 10011010
-f 10011011
-" 1001110
-T 1001111
-u 101000
-m 101001
-s 10101
-d 10110
 LF 10111
 SP 110
-a 1110
+" 1001110
+, 010010
+. 1001100
+; 0100001
 ? 111100000
+A 0100010
 B 111100001
 H 111100010
 I 111100011
-p 1111001
-c 111101
 M 1111100
+T 1001111
+W 0100000
+a 1110
+b 010011
+c 111101
+d 10110
+e 001
+f 10011011
 g 1111101
+h 0101
+i 01111
+k 10011010
+l 0110
+m 101001
+n 10010
+o 0000
+p 1111001
+r 0001
+s 10101
+t 1000
+u 101000
+v 0100011
 w 111111
+y 01110
